Add more instruction

+++ simulator.h
@@ -6,7 +6,7 @@
 u32 calc_inst_pointer_address(CPU *cpu);
 u32 calc_stack_pointer_address(CPU *cpu);
 
-s32 get_data_from_register(CPU *cpu, Register_Access *src_reg);
+u16 get_data_from_register(CPU *cpu, Register_Access *src_reg);
 
 void load_executable(CPU *cpu, char *filename);
 void boot(CPU *cpu);
